+++ CloudHarvestCoreTasks/chains/base.py
@@ -1,12 +1,12 @@
 from CloudHarvestCorePluginManager import register_definition
 from CloudHarvestCoreTasks.tasks.base import BaseTask, TaskStatusCodes
-from CloudHarvestCoreTasks.exceptions import TaskTerminationException
+from CloudHarvestCoreTasks.exceptions import TaskTerminationError
 
 from datetime import datetime, timezone
 from typing import List, Dict, Any, Generator
 from logging import getLogger
 
-from CloudHarvestCoreTasks.exceptions import TaskChainException
+from CloudHarvestCoreTasks.exceptions import TaskChainError
 
 logger = getLogger('harvest')
 
@@ -97,12 +97,11 @@ class BaseTaskChain(List[BaseTask]):
         self.position = 0
 
         self.agent = None       # populated by agent
+        self.errors = None
         self.start = None
         self.end = None
 
-        self.meta = {
-            'Errors': []
-        }
+        self.meta = {}
 
         # When set, the TaskChain will send results to this silo when it completes or errors. Can only be set after the
         # TaskChain has been instantiated. This is to prevent users from setting the silo in the template, thus sending
@@ -121,8 +120,8 @@ class BaseTaskChain(List[BaseTask]):
                 self.update_status_client = silo.connect()
                 self.update_status()
 
-        except Exception as ex:
-            logger.error(f'{self.redis_name} failed to connect to `harvest-tasks` silo: %s', ex)
+        except BaseException as ex:
+            raise TaskChainError(f'{self.redis_name} failed to connect to `harvest-tasks` silo', ex) from ex
 
     def __enter__(self) -> 'BaseTaskChain':
         """
@@ -147,22 +146,6 @@ class BaseTaskChain(List[BaseTask]):
         """
 
         return None
-
-    @property
-    def errors(self) -> List[dict]:
-        """
-        Returns a list of errors that occurred during the task chain.
-        """
-
-        errors = []
-        for task in self:
-            if task.meta.get('Errors'):
-                errors.append({f'{self.index(task)}-{task.name}': task.meta['Errors']})
-
-        if self.meta.get('Errors'):
-            errors.append({'TaskChain': self.meta['Errors']})
-
-        return errors
 
     @property
     def percent(self) -> float:
@@ -446,7 +429,7 @@ class BaseTaskChain(List[BaseTask]):
         position = self.find_task_position_by_name(task_name)
 
         if position < self.position:
-            raise TaskChainException(self, 'Cannot insert a task before the current task.')
+            raise TaskChainError(self, 'Cannot insert a task before the current task.')
 
         else:
             self.task_templates.insert(position - 1, new_task_configuration)
@@ -490,7 +473,7 @@ class BaseTaskChain(List[BaseTask]):
 
         return self
 
-    def on_error(self, ex: Exception) -> 'BaseTaskChain':
+    def on_error(self, ex: BaseException) -> 'BaseTaskChain':
         """
         Method to run when the task chain errors.
         This method may be overridden in subclasses to provide specific error handling logic.
@@ -500,7 +483,6 @@ class BaseTaskChain(List[BaseTask]):
         """
 
         self.status = TaskStatusCodes.error
-        self.meta['Errors'].append(str(ex.args))
 
         if self.pool.queue_size:
             self.pool.terminate()
@@ -535,20 +517,23 @@ class BaseTaskChain(List[BaseTask]):
             from CloudHarvestCoreTasks.silos import get_silo
             from json import dumps
             silo = get_silo(self.results_silo)
-
+            from redis import StrictRedis
             try:
-                client = silo.connect()
+                client: StrictRedis = silo.connect()
+
                 client.hset(
                     name=self.redis_name,
-                    key='result',
-                    value=dumps(results, default=str)
+                    mapping={
+                        'errors': dumps(self.errors),
+                        'result': dumps(results, default=str),
+                    }
                 )
 
                 # Sets an expiration to retrieve the results
                 client.expire(self.id, 3600)
 
-            except Exception as ex:
-                logger.error(f'{self.redis_name}: Error storing task chain results in silo {self.results_silo}: {ex}')
+            except BaseException as ex:
+                raise TaskChainError(self, f'Error storing task chain results in silo {self.results_silo}: {ex}') from ex
 
             else:
                 logger.debug(f'{self.redis_name}: Stored task chain results in silo {self.results_silo}')
@@ -640,7 +625,7 @@ class BaseTaskChain(List[BaseTask]):
 
                 # Check for termination
                 if self.status == TaskStatusCodes.terminating:
-                    raise TaskTerminationException('Task chain was instructed to terminate.')
+                    raise TaskTerminationError('Task chain was instructed to terminate.')
 
                 # Hold within the loop if there are outstanding pool tasks because the async task might have an
                 # on_* directive which needs to be added and processed. By waiting here, we ensure that the task chain
@@ -654,7 +639,7 @@ class BaseTaskChain(List[BaseTask]):
             if self.pool.queue_size > 0:
                 self.pool.wait_until_complete()
 
-        except Exception as ex:
+        except BaseException as ex:
             self.on_error(ex)
 
         finally:
@@ -729,9 +714,8 @@ class BaseTaskChain(List[BaseTask]):
                 self.update_status_client.hset(name=self.redis_name, mapping=format_hset(self.redis_struct()))
                 self.update_status_client.expire(name=self.redis_name, time=3600)
 
-            except Exception as ex:
-                logger.error(f'{self.redis_name} failed to update task chain status: %s', ex)
-                self.meta['Errors'].append(f'Error updating task chain status: {ex}')
+            except BaseException as ex:
+                raise TaskChainError(self, f'Error updating task chain status: {ex}') from ex
 
 class BaseTaskPool:
     """